+++ app/ui/Sections/Files/FilesList/components/FileDetails.jsx
