+++ app/ui/Sections/Printers/PrinterAdd/PrinterAdd.jsx
@@ -8,7 +8,7 @@ import SearchNetworkPrinters from './graphql/SearchNetworkPrinters.graphql'
 import PrinterName from './graphql/PrinterName.graphql'
 import AddPrinter from './graphql/AddPrinter.graphql'
 
-import classes from '../../../common/Common.module.less'
+import classes from 'common/Common.module.less'
 
 export default function PrinterAdd () {
   const [delay] = useState(1000)
