+++ app/ui/Sections/Printers/PrintersList/PrintersList.jsx
@@ -8,7 +8,7 @@ import { PageHeader } from '@ant-design/pro-layout'
 import Printers from './graphql/Printers.graphql'
 import NetworkPrinters from './NetworkPrinters'
 
-import classes from '../../../common/Common.module.less'
+import classes from 'common/Common.module.less'
 
 export default function PrintersList () {
   const { loading, error, data: printersData } = useQuery(Printers)
