+++ app/ui/graphql/CreateLabel.graphql
